refactor: rename gistview to base view
