+++ messaginginapp/src/main/java/io/customer/messaginginapp/gist/presentation/GistModalActivity.kt
@@ -20,6 +20,7 @@ import io.customer.messaginginapp.gist.utilities.ModalAnimationUtil
 import io.customer.messaginginapp.state.InAppMessagingAction
 import io.customer.messaginginapp.state.InAppMessagingState
 import io.customer.messaginginapp.state.MessageState
+import io.customer.messaginginapp.ui.InAppMessageViewEventsListener
 import io.customer.sdk.core.di.SDKComponent
 import io.customer.sdk.tracking.TrackableScreen
 import kotlinx.coroutines.Job
@@ -27,7 +28,7 @@ import kotlinx.coroutines.Job
 const val GIST_MESSAGE_INTENT: String = "GIST_MESSAGE"
 const val GIST_MODAL_POSITION_INTENT: String = "GIST_MODAL_POSITION"
 
-class GistModalActivity : AppCompatActivity(), GistViewListener, TrackableScreen {
+class GistModalActivity : AppCompatActivity(), InAppMessageViewEventsListener, TrackableScreen {
     private lateinit var binding: ActivityGistBinding
     private var elapsedTimer: ElapsedTimer = ElapsedTimer()
     private val inAppMessagingManager = SDKComponent.inAppMessagingManager
@@ -193,7 +194,7 @@ class GistModalActivity : AppCompatActivity(), GistViewListener, TrackableScreen
         finish()
     }
 
-    override fun onGistViewSizeChanged(width: Int, height: Int) {
+    override fun onViewSizeChanged(width: Int, height: Int) {
         logger.debug("GistModelActivity Size changed: $width x $height")
         val params = binding.gistView.layoutParams
         params.height = height
