+++ messaginginapp/src/main/java/io/customer/messaginginapp/ui/InAppMessageViewEventsListener.kt
@@ -1,0 +1,8 @@
+package io.customer.messaginginapp.ui
+
+/**
+ * Listener interface to send events from the InAppMessageHostView to the host activity or fragment.
+ */
+internal interface InAppMessageViewEventsListener {
+    fun onViewSizeChanged(width: Int, height: Int) {}
+}
